requirements.txt
openenv-core>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
sqlglot>=20.0.0
psycopg2-binary>=2.9.0
openai>=1.0.0
requests>=2.31.0
pytest>=7.4.0
testing.postgresql>=1.3.0
python-dotenv>=1.0.0
